+++ spec/recipes/default_spec.rb
@@ -4,11 +4,11 @@ require_relative '../spec_helper'
 
 describe 'chef-dk::default' do
   let(:platform) { { platform: 'ubuntu', version: '14.04' } }
-  let(:overrides) { {} }
+  %i(version source global_shell_init).each { |a| let(a) { nil } }
   let(:runner) do
     ChefSpec::SoloRunner.new(platform) do |node|
-      overrides.each do |k, v|
-        node.default['chef_dk'][k] = v
+      %i(version source global_shell_init).each do |a|
+        node.default['chef_dk'][a] = send(a) unless send(a).nil?
       end
     end
   end
@@ -16,34 +16,25 @@ describe 'chef-dk::default' do
 
   context 'with default attributes' do
     it 'installs the latest version of the Chef-DK' do
-      expect(chef_run).to install_chef_dk('chef_dk').with(version: 'latest')
+      expect(chef_run).to install_chef_dk('chef_dk')
+        .with(version: nil, source: :direct, global_shell_init: false)
     end
   end
 
   context 'an overridden `version` attribute' do
-    let(:overrides) { { version: '1.2.3-4' } }
+    let(:version) { '1.2.3-4' }
 
     it 'installs the specified version of the Chef-DK' do
       expect(chef_run).to install_chef_dk('chef_dk').with(version: '1.2.3-4')
     end
   end
 
-  context 'an overridden `package_url` attribute' do
-    let(:overrides) { { package_url: 'http://example.com/pkg.pkg' } }
+  context 'an overridden `source` attribute' do
+    let(:source) { 'http://example.com/pkg.pkg' }
 
     it 'installs from the desired package URL' do
       expect(chef_run).to install_chef_dk('chef_dk')
-        .with(package_url: 'http://example.com/pkg.pkg')
-    end
-  end
-
-  context 'overridden `version` and `package_url` attributes' do
-    let(:overrides) do
-      { version: '1.2.3-4', package_url: 'http://example.com/pkg.pkg' }
-    end
-
-    it 'raises an exception' do
-      expect { chef_run }.to raise_exception(Chef::Exceptions::ValidationFailed)
+        .with(source: 'http://example.com/pkg.pkg')
     end
   end
 end